Refactor global Doctrine listeners to targeted entity listeners
Replace the single global SecurityTokenPaymentLinker (registered for all
entities via #[AsDoctrineListener]) with two focused entity listeners:

- SecurityTokenPaymentLinker: #[AsEntityListener] for SecurityToken
  prePersist/preUpdate only — Doctrine no longer invokes it for every
  persist/update in the system.

- PaymentSecurityTokenRemover: #[AsEntityListener] for Payment
  preRemove only — replaces the preRemove branch that lived in the old
  global listener.

With entity listeners Doctrine passes the typed entity as the first
argument, so the instanceof guards and getObject() calls are removed.
Tests updated to match the new (entity, args) call signatures; the
"ignores wrong entity type" cases are deleted as they are no longer
applicable.
